# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tasks": {
      "post": {
        "description": "Creates a new task for the authenticated user",
        "operationId": "TasksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": ["title must be longer than or equal to 3 characters"],
                    "error": "Bad Request",
                    "errorCode": "VALIDATION_FAILED",
                    "timestamp": "2025-12-13T10:30:00.000Z",
                    "path": "/api/v1/tasks",
                    "correlationId": "550e8400-e29b-41d4-a716-446655440000"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new task",
        "tags": ["tasks"]
      },
      "get": {
        "description": "Retrieves a paginated list of tasks with optional filtering and sorting. Supports filtering by status, priority, and search term.",
        "operationId": "TasksController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 10, max: 100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by task status",
            "schema": {
              "enum": ["TODO", "IN_PROGRESS", "COMPLETED", "CANCELLED"],
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "description": "Filter by task priority",
            "schema": {
              "enum": ["LOW", "MEDIUM", "HIGH"],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in task title and description",
            "schema": {
              "example": "meeting",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by (default: CREATED_AT)",
            "schema": {
              "enum": ["CREATED_AT", "UPDATED_AT", "DUE_DATE", "TITLE", "PRIORITY", "STATUS"],
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order (default: DESC)",
            "schema": {
              "enum": ["ASC", "DESC"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTasksResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all tasks",
        "tags": ["tasks"]
      }
    },
...
    "/tasks/{id}": {
      "get": {
        "description": "Returns task only if user is the owner or an admin",
        "operationId": "TasksV2Controller_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task ID (CUID format)",
            "schema": {
              "example": "cm4abc123xyz456def789ghi",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found and user has permission to access it",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token missing or invalid"
          },
          "403": {
            "description": "Access denied - user is not the task owner or admin",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "You do not have permission to access this task",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Task with ID cm4abc123xyz456def789ghi not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get task by ID (v2 - with permission checks)",
        "tags": ["tasks-v2"]
      },
      "patch": {
        "description": "Updates an existing task. Only the task owner can update their tasks.",
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task CUID",
            "schema": {
              "example": "cmixpvpir0001p9yp5xq8r7ks",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Invalid CUID format: invalid-id",
                    "error": "Bad Request",
                    "errorCode": "VALIDATION_INVALID_CUID",
                    "timestamp": "2025-12-13T10:30:00.000Z",
                    "path": "/api/v1/tasks/invalid-id",
                    "correlationId": "550e8400-e29b-41d4-a716-446655440000"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Cannot update other user's task",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "You do not have permission to access task: cmixpvpir0001p9yp5xq8r7ks",
                    "error": "Forbidden",
                    "errorCode": "TASK_FORBIDDEN",
                    "timestamp": "2025-12-13T10:30:00.000Z",
                    "path": "/api/v1/tasks/cmixpvpir0001p9yp5xq8r7ks",
                    "correlationId": "550e8400-e29b-41d4-a716-446655440000"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Task with ID 'cmixpvpir0001p9yp5xq8r7ks' not found",
                    "error": "Not Found",
                    "errorCode": "TASK_NOT_FOUND",
                    "timestamp": "2025-12-13T10:30:00.000Z",
                    "path": "/api/v1/tasks/cmixpvpir0001p9yp5xq8r7ks",
                    "correlationId": "550e8400-e29b-41d4-a716-446655440000"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a task",
        "tags": ["tasks"]
      },
      "delete": {
        "description": "Soft deletes a task. The task will be marked as deleted but retained in the database.",
        "operationId": "TasksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task CUID",
            "schema": {
              "example": "cmixpvpir0001p9yp5xq8r7ks",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task successfully deleted"
          },
          "403": {
            "description": "Cannot delete other user's task",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "You do not have permission to access task: cmixpvpir0001p9yp5xq8r7ks",
                    "error": "Forbidden",
                    "errorCode": "TASK_FORBIDDEN",
                    "timestamp": "2025-12-13T10:30:00.000Z",
                    "path": "/api/v1/tasks/cmixpvpir0001p9yp5xq8r7ks",
                    "correlationId": "550e8400-e29b-41d4-a716-446655440000"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Task with ID 'cmixpvpir0001p9yp5xq8r7ks' not found",
                    "error": "Not Found",
                    "errorCode": "TASK_NOT_FOUND",
                    "timestamp": "2025-12-13T10:30:00.000Z",
                    "path": "/api/v1/tasks/cmixpvpir0001p9yp5xq8r7ks",
                    "correlationId": "550e8400-e29b-41d4-a716-446655440000"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a task",
        "tags": ["tasks"]
      }
    },
...
    "/tasks/next-due-date": {
      "get": {
        "description": "Returns the task with the soonest due date that has not passed yet. Only includes tasks with status TODO or IN_PROGRESS (excludes COMPLETED and CANCELLED). Returns null if no upcoming active tasks exist.",
        "operationId": "TasksV2Controller_getNextDueTask",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Next due task found (or null if none)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token missing or invalid"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get task with nearest upcoming due date",
        "tags": ["tasks-v2"]
      }
    },
...
  },
  "info": {
    "title": "NestJS API",
    "description": "A production-ready REST API built with NestJS",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/yourusername/nestjs-api-skeleton",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the task",
            "example": "Implement authentication",
            "minLength": 3,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the task",
            "example": "Add JWT authentication to the API",
            "maxLength": 1000
          },
          "status": {
            "type": "string",
            "description": "Status of the task",
            "enum": ["TODO", "IN_PROGRESS", "COMPLETED", "CANCELLED"],
            "example": "TODO",
            "default": "TODO"
          },
          "priority": {
            "type": "string",
            "description": "Priority level of the task",
            "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"],
            "example": "MEDIUM",
            "default": "MEDIUM"
          },
          "dueDate": {
            "type": "string",
            "description": "Due date for the task in ISO 8601 date string format",
            "example": "2025-12-31T23:59:59.000Z",
            "format": "date-time"
          }
        },
        "required": ["title"]
      },
      "TaskResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the task",
            "example": "clxyz123456789"
          },
          "title": {
            "type": "string",
            "description": "Title of the task",
            "example": "Complete project documentation"
          },
          "description": {
            "type": "object",
            "description": "Detailed description of the task",
            "example": "Write comprehensive documentation for all API endpoints"
          },
          "status": {
            "type": "string",
            "description": "Current status of the task",
            "enum": ["TODO", "IN_PROGRESS", "COMPLETED", "CANCELLED"],
            "example": "IN_PROGRESS"
          },
          "priority": {
            "type": "string",
            "description": "Priority level of the task",
            "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"],
            "example": "HIGH"
          },
          "dueDate": {
            "format": "date-time",
            "type": "string",
            "description": "Due date of the task",
            "example": "2025-12-31T23:59:59.000Z"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the task was completed",
            "example": "2025-12-10T15:30:00.000Z"
          },
          "userId": {
            "type": "string",
            "description": "User ID who owns the task",
            "example": "cluserid123456"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the task was created",
            "example": "2025-12-10T08:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the task was last updated",
            "example": "2025-12-10T09:15:00.000Z"
          }
        },
        "required": ["id", "title", "status", "priority", "userId", "createdAt", "updatedAt"]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total number of items",
            "example": 42
          },
          "page": {
            "type": "number",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Number of items per page",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of pages",
            "example": 5
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Whether there is a next page",
            "example": true
          },
          "hasPrevPage": {
            "type": "boolean",
            "description": "Whether there is a previous page",
            "example": false
          }
        },
        "required": ["total", "page", "limit", "totalPages", "hasNextPage", "hasPrevPage"]
      },
      "PaginatedTasksResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Array of tasks",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskResponseDto"
            }
          },
          "meta": {
            "description": "Pagination metadata",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationMetaDto"
              }
            ]
          }
        },
        "required": ["data", "meta"]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the task",
            "example": "Implement authentication",
            "minLength": 3,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the task",
            "example": "Add JWT authentication to the API",
            "maxLength": 1000
          },
          "status": {
            "type": "string",
            "description": "Status of the task",
            "enum": ["TODO", "IN_PROGRESS", "COMPLETED", "CANCELLED"],
            "example": "IN_PROGRESS",
            "default": "TODO"
          },
          "priority": {
            "type": "string",
            "description": "Priority level of the task",
            "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"],
            "example": "MEDIUM",
            "default": "MEDIUM"
          },
          "dueDate": {
            "type": "string",
            "description": "Due date for the task in ISO 8601 date string format",
            "example": "2025-12-31T23:59:59.000Z",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "description": "Timestamp when the task was completed in ISO 8601 format",
            "example": "2025-12-10T15:30:00.000Z",
            "format": "date-time"
          }
        }
      }
    }
  }
}